backend/requirements.txt
fastapi
uvicorn[standard]
pymupdf
python-multipart
python-jose
passlib[bcrypt]